backend/requirements.txt
python-dotenv
requests
gtfs-realtime-bindings
supabase
